restructured base layout for pages
