+++ src/pages/Detail.tsx
@@ -1,8 +1,7 @@
 import { useParams } from 'react-router-dom';
-import { Page } from '../components/Page';
 import { BackButton } from '../components/buttons/BackButton';
 import { Description } from '../components/detail/description/Description';
-import Header from '../components/header/Header';
+import { BaseLayout } from '../components/layout/BaseLayout';
 import LoadingSpinner from '../components/loadingSpinner/LoadingSpinner';
 import { Splash } from '../components/splash/Splash';
 import useFetchTopic from '../hooks/sections/useFetchTopic';
@@ -12,17 +11,16 @@ export const Detail = () => {
   const { data: topic, isLoading } = useFetchTopic(parseInt(detailId));
 
   return (
-    <Page>
+    <BaseLayout>
       {isLoading && <LoadingSpinner />}
       {!isLoading && (
         <>
-          <Header />
           <Splash sectionBgColor={topic!.bgColor} title={topic!.title} />
           <Description description={topic!.description} />
           {/*<MoreContents data={topics} canReplace={true} />*/}
           <BackButton />
         </>
       )}
-    </Page>
+    </BaseLayout>
   );
 };
