+++ src/pages/Main.tsx
@@ -1,5 +1,4 @@
-import { Page } from '../components/Page';
-import Header from '../components/header/Header';
+import { BaseLayout } from '../components/layout/BaseLayout';
 import LoadingSpinner from '../components/loadingSpinner/LoadingSpinner';
 import GuideSection from '../components/main/guideSection/GuideSection';
 import { RectorSection } from '../components/main/rector/RectorSection';
@@ -10,17 +9,16 @@ const Main = () => {
   const { data: landing, isLoading } = useFetchLanding();
 
   return (
-    <Page>
+    <BaseLayout>
       {isLoading && <LoadingSpinner />}
       {!isLoading && (
         <>
-          <Header />
           <Splash />
           <GuideSection />
           <RectorSection />
         </>
       )}
-    </Page>
+    </BaseLayout>
   );
 };
